Fix scyjava problems on Windows
The version of Maven packaged by conda includes both mvn and mvn.cmd;
only mvn.cmd works when invoked from Python. Invoking mvn yields:

  FileNotFoundError: [WinError 2] The system cannot find the file specified

Then, once mvn.cmd is used, the output capture has Windows-style line
breaks ('\r\n') instead of Unix-style ('\n'). So we adjust for that.

Then, when parsing the path, Windows uses backslash rather than forward
slash, so we cannot rely on '/' as the leading path character. Instead,
we look for the colon and jump ahead two characters. This assumes only
one space after the colon, but Maven appears to be reliable about that.

Then, when attempting to link to the JVM, we need jvm.dll on the PATH
too, or else the linkage fails with:

  ImportError: DLL load failed: The specified module could not be found.

The jvm.dll resides in a subfolder called server (for older Javas it's
jre\bin\server; for newer ones it's bin\server). So we look for jvm.dll
in those places, and if found, add the containing folder to the PATH.

See also imagej/pyimagej#43.