Improve JVM shutdown
We are overriding JPype's builtin exit handling
with our own so that we can provide an extensible callback
mechanism for downstream uses. In this way consumers can ensure
that specific things happen before the JVM shutsdown.

We also catch ^C keyboard interrupt, so termination
occurs even with the interactive shell.