Run tests cross-platform
This should help us to catch platform-specific problems.

Note that the individual tests are run explicitly instead of in a
for-loop, due to Windows running in powershell on default and not having
an easy consistent syntax that also ensures appropriate error code
propagation. This is detrimental to extensiblilty and should be resolved
if more tests are added.