Fix java version check failure on Windows
scyjava's version check fails with:
TypeError: argument of type 'WindowsPath' is not iterable
when on Windows and using Python <=3.7. Is seems that Pathlib.path only
became iterable with version 3.8. Casting the 'WindowsPath' as a string
resolves this issue.

I tested this with Python versions 3.6.13, 3.7.11, 3.8.12 and 3.9.7.
All tests pass on Windows and Linux for the same Python version range.