Fix mangled text data columns
When creating a pandas DataFrame form the extracted
org.scijava.table.Table contents, the resulting pandas
DataFrame containes char arrays of the string/text column entries. The issue
is caused when converting the List of java objects into
the pandas DataFrame. Converting the contents of the the List
to Python with 'to_python' before creating the DataFrame
fixes this bug.