Introduce is_version_at_least function
And deprecate compare_version.

The name compare_version is unclear, and the behavior is surprising (to
me): the given version must be *strictly greater than* some stated
minimum. But our primary use case for this query is "I need to know if
library X is new enough to call function Y, so I'll check whether
library X is at least version Z, which was the first version to include
function Y." The new function name and behavior is_version_at_least is
more intuitive for that. If you need to know strictly greater than, you
can check for is_version_at_least together with inequality.