Only dispose AWT windows if AWT has started
On macOS, even jimport-ing an AWT-related class can trigger the startup
of Java's AWT subsystem, which leads to a deadlock when not running
headless or using setupGuiEnvironment. To avoid this conundrum, we use a
heuristic to check for AWT: are there any running threads with the AWT-
prefix? If not, don't bother trying to clean up any AWT things.

At least on my macOS system, this change fixes deadlocks on shutdown,
so I'm hopeful that it will fix some scenarios on macOS CI as well.