Fix locked Java buffer copy for array conversions
`memoryview` of Java arrays (jarr) exposes the jarr's
buffer protocol and copies the buffer into a page
which then is referenced locked by `memoryview` and
things that are created from it (e.g. numpy.frombuffer). Changing
the underlying jarr's memory and requesting a new `memoryview`
does not return an updated view with the jarr's data change as
we expect, but instead the initial `memoryview` data state. This is bad.
In order to get `memoryview` to reflect the correct jarr data changes
you need to dereference (e.g. del or set to None) the `memoryview`
and objects that depend on it (e.g. a numpy array created from the
view).

This patch resolves this issue by obtaining the buffer
that `memoryview` is getting and wrapping the buffer
with numpy as we want. We can then safely release the buffer
(preserving memory) and return a copy to the user. If
_jarray_to_ndarray is called again on the same Java array
a new buffer is obtained, wrapped with numpy and released. The end
result is an updated view of the Java array upon request.