Fix bug in optional dependency import
There were two conflicting assumptions in the codebase:

1. If the library is not found, a falsy value is returned; and
2. If the library is not found, an exception is raised.

When initializing converters, it's looking for the falsy value,
i.e. just a simple test of whether the dependency is available.
But when importing the dependency in a function that actually uses it,
an exception is preferred, to ensure the import never ends up as None.

This commit fulfills both scenarios by adding a boolean required flag
to the respective import functions. If a gentle check for availability
is preferred, one can pass required=False to obtain that behavior now.