conda: set openjdk version ranges more carefully
Certainly we require at least OpenJDK 8, although on conda-forge,
8.0.112 is the oldest anyway, so this rule has no effect in practice.

For the developer environment, for now we restrict to OpenJDK 11
maximum, not OpenJDK 17, because we want CI to use OpenJDK 11
(the scijava-tables library has a bug with OpenJDK 17 as of this
writing). In the long term, restricting OpenJDK in this way is the
wrong thing to do, but it's tolerable for now to get tests passing.