Be more conservative about overriding sys.stdout
Unfortunately, in Jupyter notebooks (at least with ipykernel 6.x),
there is already stdout redirection going on, to support printing
outputs into the browser below the cell. And for some reason, even
though the existing sys.stdout we are overriding *is* an ipykernel
output object, it somehow stops being able to yield output below the
cells after we replace it. Maybe Jupyter replaces the stdout more than
once, and we have a stale object instead of the latest/current one?

Regardless, we want to avoid overriding stdout permanently, in favor
of using the contextlib.redirect_stdout function to override it only
during script evaluation.

It's not clear that this function will actually do the right thing
if/when multiple Python scripts are executed concurrently, since
effectively there will be nested `with redirect_stdout` blocks in
that scenario. But this change should make Jupyter notebooks behave
better again in the vastly most common case where SciJava Python
scripts are not being executed (because why would you execute a
SciJava Python script from inside a Python kernel Jupyter notebook,
when you can just run the code directly in the cell?).