Add convenience functions for checking JVM config
After the JVM has started, it is nice to be able to easily check
characteristics of the JVM from the Runtime API, including current and
maximum heap sizes, as well as number of available processors. (It is
also nice to be able to easily split infinitives in a commit message.)

Relatedly, it can be nice to be able to call the garbage collector
directly.

Note that we do not expose the Runtime.getRuntime().freeMemory() method,
because it can be confusing: one would naively expect freeMemory() to
equal maxMemory() - usedMemory(), but in actuality it represents the
amount of free memory within the bounds of the memory already reserved
by the JVM *at the moment*. So it is typically not a very useful number.