Bump openjdk version to 11
This commit resolves an unusual bug where pytest
fails to collect the results of the jep test, resulting
in failing CI jobs. I was able to reproduce this locally,
sometimes, with openjdk=8. There must be a race condition involved
with the bug as it doesn't always fail. The crash is a seg fault.