Be less aggressive with source code detection
When invoking scyjava.inspect functions, they can optionally report the
source URL at the top, before printing the members. But this only works
if scijava-search is on the classpath. Let's let the source flag default
to None, in which case it swallows source code URL detection failures
gracefully, to make the common case of scijava-search not being
available work without hassle.

And let's have the various inspect functions accept static and source
boolean flags, which get passed along to the internal _print_data
routine, as was previously the case when they were partial functions.